otp: Move OTP confirm (login) view/form/template to otp_views
Now all OTP-related views, forms and templates are in the same place.
