+++ itou/utils/context_processors.py
@@ -19,7 +19,7 @@ def matomo(request):
 
 
 def active_announcement_campaign(request):
-    if request.user and request.user.is_authenticated and not request.path.startswith("/login/verify"):
+    if request.user and request.user.is_authenticated and not request.path.startswith("/otp/verify"):
         if campaign := get_cached_active_announcement():
             return {
                 "display_campaign_announce": True,
