+++ itou/utils/perms/middleware.py
@@ -199,7 +199,7 @@ class ItouCurrentOrganizationMiddleware:
             and user.kind == UserKind.ITOU_STAFF
             and not user.is_verified()
         ):
-            login_verify_otp_url = reverse("login:verify_otp")
+            login_verify_otp_url = reverse("otp_views:verify_otp")
             login_with_backup_code_url = reverse("otp_views:login_with_backup_code")
             if user_has_device(user):
                 if request.path not in (login_verify_otp_url, login_with_backup_code_url):
