+++ itou/www/login/urls.py
@@ -8,7 +8,6 @@ app_name = "login"
 
 urlpatterns = [
     path("existing", views.ExistingUserLoginView.as_view(), name="existing_user"),
-    path("verify", views.VerifyOTPView.as_view(), name="verify_otp"),
     path("demo", views.demo_login_view, name="demo"),
     # Retro compatibility url
     path("job_seeker", views.PreLoginView.as_view(), name="job_seeker"),  # FIXME use redirect instead ?
