+++ itou/www/otp_views/forms.py
@@ -1,6 +1,7 @@
 from base64 import b32encode
 
 from django import forms
+from django_otp import match_token
 
 from itou.www.otp_views.enums import DeviceType
 
@@ -55,3 +56,36 @@ class LoginWithBackupCodeForm(forms.Form):
             self.add_error("code", "Le code de récupération n’est pas correct.")
 
         return cleaned_data
+
+
+class VerifyOTPForm(forms.Form):
+    otp_token = forms.CharField(
+        required=True,
+        label="Entrez le code de validation unique (OTP)",
+        help_text=(
+            "Code à 6 chiffres généré par votre application mobile "
+            "ou votre gestionnaire de mot de passe sur votre ordinateur"
+        ),
+    )
+
+    otp_token.widget.attrs.update(
+        {
+            "max_length": 6,
+            "autocomplete": "one-time-code",
+            "autofocus": True,
+        }
+    )
+
+    def __init__(self, *args, user, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.user = user
+
+    def clean_otp_token(self):
+        otp_token = self.cleaned_data.get("otp_token")
+
+        device = match_token(self.user, otp_token)
+        if device is None:
+            raise forms.ValidationError("Le code de validation unique (OTP) n’est pas correct.")
+        self.user.otp_device = device
+
+        return otp_token
