+++ tests/utils/tests.py
@@ -1622,7 +1622,7 @@ def test_active_announcement_campaign_context_processor(client, empty_active_ann
     assert response.context["display_campaign_announce"] is True
 
     client.force_login(ItouStaffFactory())
-    response = client.get(reverse("login:verify_otp") + "?next=%2Fadmin%2F")
+    response = client.get(reverse("otp_views:verify_otp") + "?next=%2Fadmin%2F")
     assert response.status_code == 200
     assert response.context["display_campaign_announce"] is False
